tls doc updates
