+++ actix-tls/Cargo.toml
@@ -77,5 +77,5 @@ tokio-rustls = { version = "0.23", features = ["dangerous_configuration"] }
 trust-dns-resolver = "0.20.0"
 
 [[example]]
-name = "tcp-rustls"
+name = "accept-rustls"
 required-features = ["accept", "rustls"]
